chore(Settings): Review the settings to have dev, test, production in different files
