+++ .env.dist
@@ -15,3 +15,6 @@ AWS_USER_ARN=
 AWS_APP_ROLE_ARN=
 AWS_PERMISSIONS_BOUNDARY_POLICY_ARN=
 WORKSPACE_DATASETS_BUCKET=
+
+WORKSPACE_STORAGE_ENGINE_AWS_ACCESS_KEY_ID=
+WORKSPACE_STORAGE_ENGINE_AWS_SECRET_ACCESS_KEY=
