+++ docker-entrypoint.sh
@@ -47,15 +47,17 @@ case "$command" in
   ;;
 "test")
   wait-for-it db:5432
-  export DEBUG=false
+  export DJANGO_SETTINGS_MODULE=config.settings.test
   python manage.py makemigrations --check
-  python manage.py test --settings=config.settings_test $arguments
+  python manage.py test $arguments
   ;;
 "coveraged-test")
   wait-for-it db:5432
-  export DEBUG=false
-  python manage.py makemigrations --settings=config.settings_test --check
-  coverage run manage.py test --settings=config.settings_test $arguments
+  export DJANGO_SETTINGS_MODULE=config.settings.test
+  python manage.py makemigrations --check
+  coverage run manage.py test $arguments
+  python manage.py makemigrations --check
+  coverage run manage.py test  $arguments
   coverage report --skip-empty --fail-under=80
   ;;
 "manage")
