+++ hexa/pipelines/utils.py
@@ -1,11 +1,11 @@
 import datetime
 
+from django.conf import settings
 from django.utils import timezone
 from django.utils.crypto import get_random_string
 from django.utils.translation import gettext_lazy, override
 from slugify import slugify
 
-from config import settings
 from hexa.core.utils import send_mail
 
 from .models import PipelineRun, PipelineRunTrigger
