+++ hexa/workspaces/utils.py
@@ -1,8 +1,8 @@
 from urllib.parse import urlencode
 
+from django.conf import settings
 from django.utils.translation import gettext_lazy
 
-from config import settings
 from hexa.core.utils import send_mail
 
 from .models import WorkspaceInvitation
